refactor: useTVLOvertime to use react-query
