+++ src/pages/Overview/Overview.tsx
@@ -6,8 +6,7 @@ import { Column, Row } from "../../components/Layout"
 import { Title } from "../../components/Title"
 import { Chart } from "../../components/Chart/Chart"
 
-import { useTVLOverTime } from "../../hooks/api/useTVLOverTime"
-import { useTVCOverTime } from "../../hooks/api/stats"
+import { useTVCOverTime, useTVLOverTime } from "../../hooks/api/stats"
 
 import styles from "./Overview.module.scss"
 import APYChart from "../../components/APYChart/APYChart"
@@ -20,12 +19,8 @@ type ChartDataPoint = {
 }
 
 export const OverviewPage: React.FC = () => {
-  const { getTVLOverTime, data: tvlData } = useTVLOverTime()
+  const { data: tvlData } = useTVLOverTime()
   const { data: tvcData } = useTVCOverTime()
-
-  useEffect(() => {
-    getTVLOverTime()
-  }, [getTVLOverTime])
 
   const chartsData = useMemo(() => {
     if (!tvlData || !tvcData) return
